Merged PR 8436: SECURITY PATCH .azurepipelines: Do not run ARM Pipelines
Disables the ARM Pipelines for the SECURE repos as these pipelines rely
on a Self-Hosted Agent that is not available in the SECURE repos.
Access could be provided, but the decision was made not to intermingle
the SECURE and public repos on the same open source agent.

----
#### AI description  (iteration 1)
#### PR Classification
Security patch

#### PR Summary
This pull request disables the execution of ARM pipelines in the Azure Pipelines configuration files to address security concerns.
- `.azurepipelines/Ubuntu-GCC5.yml`: Commented out ARM pipeline targets.
- `.azurepipelines/Windows-VS.yml`: Commented out ARM pipeline targets.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->